LeftSideMovementForward/Backward
I wanted to create fluidity in the movements, so I had my brother video
tape me. From that I made 7 pictures that would create a movement,
forward and backward, depending on whether you are moving through them
forwards or backwards. When looped the movement would continue. It would
take about 1 second and a half a second to get the full movement.
I am not sure if this could work in the game, but I guess this would be
my sprite for movement on the left side. I want to now edit these
pictures, not putting myself as the character, but changing the face and
adding muscles etc.
Tell me what you all think of the movement and if this can work!